requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.23.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.4.2
pydantic-settings==2.0.3
python-dotenv==1.2.2
python-multipart==0.0.6
requests==2.32.5
slowapi==0.1.9
sentry-sdk[fastapi]==1.40.6
# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0
resend==0.7.0
httpx==0.27.0
